bump: electron-log
