+++ packages/insomnia/src/entry.client.tsx
@@ -1,4 +1,5 @@
 import './ui/rendererListeners';
+import './ui/log';
 
 import { startTransition, StrictMode } from 'react';
 import { hydrateRoot } from 'react-dom/client';
@@ -8,7 +9,6 @@ import { HydratedRouter } from 'react-router/dom';
 import { migrateFromLocalStorage, setSessionData, setVaultSessionData } from './account/session';
 import { getInsomniaSession, getInsomniaVaultKey, getInsomniaVaultSalt, getSkipOnboarding } from './common/constants';
 import { database } from './common/database';
-import { initializeLogging } from './common/log';
 import { settings } from './models';
 import { initNewOAuthSession } from './network/o-auth-2/get-token';
 import { init as initPlugins } from './plugins';
@@ -22,7 +22,6 @@ import { initializeSentry } from './ui/sentry';
 import { getInitialEntry } from './utils/router';
 
 initializeSentry();
-initializeLogging();
 
 try {
   window.showAlert = options => showModal(AlertModal, options);
