+++ packages/insomnia/src/ui/log.ts
@@ -1,0 +1,3 @@
+import log from 'electron-log/renderer';
+
+Object.assign(console, log.functions);
